loaded full post which user selected .Display title and  body of user selected post using id property
It has one error if you go to network tab you will see continous sending of request (infinite request) because we are upading
state from componentDidUpdate() method. we call setState in componentDidUpdate() which update component and called componentDidUpdate() method in this again setState() method will call and update the component and it goes on.