Loading FullPost using route.
In Posts.js NavLink is added so that we will send post.id parameter to url.
In FullPost.js we have use this.props.match.params.id to receive id of post send from Blog.js
also isNan is added so that when user visit NewPost link it wont receive id which is integer it will receive a text link "/new-post" in that case it will generate error in console to avoid that we use it.