for Testing with enzyme we install below packages
npm install --save enzyme
npm install --save enzyme-adapter-react-16
npm install --save react-test-renderer

we install this packages to do testing in react.

run in the console : npm test  to test the code written in NavigationItems.test.js

In this file we are testing if user authenticated there should be 3 NavigationItems and if user is not authenticated 2 NavigationItems should be there.